Hero: Resolve tall video bug
UI was fine in development, but in production had a big gap between the video and it's top and bottom extents (control bar etc) on small screens.
This was mostly resolved by removing the max-height attribute from the video element media query. The attribute is now set to 90vh in the standard video selector.